Change `lha_decoder_get_length` to return 64-bit
To keep ABI compatibility we can't really change this, but we can define
a new function that returns a 64-bit value. With some #define magic we
can invisibly redirect any callers over to use our new function instead.
We keep a compatible version of the function under the old name that
continues returning `size_t` for compatibility (on some systems, this is
a 32-bit value).

Part of #60.

+++ lib/lha_decoder.h
@@ -94,7 +94,7 @@ struct _LHADecoder {
 
 	/** Current position in the decode stream, and total length. */
 
-	size_t stream_pos, stream_length;
+	uint64_t stream_pos, stream_length;
 
 	/** Output buffer, containing decoded data not yet returned. */
